move first components to common folder
