+++ src/components/common/SearchBar.vue
