+++ src/components/course/list/lecturer/LecturerCourse.vue
@@ -38,7 +38,7 @@
 </template>
 
 <script lang="ts">
-    import ReadMore from "@/components/ReadMore.vue";
+    import ReadMore from "@/components/common/ReadMore.vue";
     import router from "@/router";
     import Course from "@/api/api_models/course_management/Course";
     import UserManagement from "@/api/UserManagement";
