+++ src/views/student/StudentCourseList.vue
@@ -15,7 +15,7 @@
 
 <script lang="ts">
     import CourseList from "@/components/course/list/common/CourseList.vue";
-    import SeachBar from "@/components/SearchBar.vue";
+    import SeachBar from "@/components/common/SearchBar.vue";
     import { ref } from "vue";
     import CourseTypeFilter from "@/components/course/list/common/CourseTypeFilter.vue";
     import { CourseType } from "@/entities/CourseType";
